refactoring following SPring Boot recommendations
