+++ ci-droid-tasks-consumer-autoconfigure/pom.xml
@@ -1,0 +1,49 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <groupId>com.societegenerale.ci-droid.tasks-consumer</groupId>
+    <artifactId>ci-droid-tasks-consumer-autoconfigure</artifactId>
+    <version>1.0.1-SNAPSHOT</version>
+    <packaging>jar</packaging>
+
+    <name>CI droid tasks consumer autoconfigure</name>
+
+    <parent>
+        <groupId>com.societegenerale.ci-droid.tasks-consumer</groupId>
+        <artifactId>ci-droid-tasks-consumer-parent</artifactId>
+        <version>1.0.1-SNAPSHOT</version>
+    </parent>
+
+
+    <description>CI-droid Tasks consumer autoconfigure</description>
+
+    <properties>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
+        <java.version>1.8</java.version>
+    </properties>
+
+    <dependencies>
+
+        <dependency>
+            <groupId>com.societegenerale.ci-droid.tasks-consumer</groupId>
+            <artifactId>ci-droid-tasks-consumer-infrastructure</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
+
+        <!-- for conditional logging with logback -->
+        <dependency>
+            <groupId>org.codehaus.janino</groupId>
+            <artifactId>janino</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>net.logstash.logback</groupId>
+            <artifactId>logstash-logback-encoder</artifactId>
+        </dependency>
+
+    </dependencies>
+
+</project>
