+++ ci-droid-tasks-consumer-autoconfigure/src/main/resources/application.yml
