+++ ci-droid-tasks-consumer-autoconfigure/src/main/resources/logback-spring.xml
