+++ ci-droid-tasks-consumer-autoconfigure/src/test/resources/application.properties
