--- ci-droid-tasks-consumer-bootstrap/pom.xml
@@ -1,71 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>com.societegenerale.ci-droid.tasks-consumer</groupId>
-        <artifactId>ci-droid-tasks-consumer</artifactId>
-        <version>1.0.1-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>ci-droid-tasks-consumer-bootstrap</artifactId>
-    <version>1.0.1-SNAPSHOT</version>
-
-    <properties>
-        <application.name.for.logstash>ci-droid-tasks-consumer</application.name.for.logstash>
-    </properties>
-
-
-    <dependencies>
-
-        <dependency>
-            <groupId>com.societegenerale.ci-droid.tasks-consumer</groupId>
-            <artifactId>ci-droid-tasks-consumer-infrastructure</artifactId>
-            <version>${project.parent.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>net.logstash.logback</groupId>
-            <artifactId>logstash-logback-encoder</artifactId>
-        </dependency>
-        <!-- for conditional logging with logback -->
-        <dependency>
-            <groupId>org.codehaus.janino</groupId>
-            <artifactId>janino</artifactId>
-        </dependency>
-
-    </dependencies>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-failsafe-plugin</artifactId>
-                <configuration>
-                    <!-- see https://github.com/spring-projects/spring-boot/issues/6254 for infos on why we need to do this -->
-                    <classesDirectory>${project.build.outputDirectory}</classesDirectory>
-                </configuration>
-            </plugin>
-
-            <plugin>
-
-                <!-- This is required so that we have more flexibility :
-
-                - the -exec jar can be used directly from command line
-                - the regular jar can be used as a dependency in projects that need it
-
-                https://stackoverflow.com/questions/48435800/using-kotlin-classes-from-jar-built-with-kotlin-maven-plugin
-
-                -->
-
-                <groupId>org.springframework.boot</groupId>
-                <artifactId>spring-boot-maven-plugin</artifactId>
-                <configuration>
-                    <classifier>exec</classifier>
-                </configuration>
-            </plugin>
-
-
-        </plugins>
-    </build>
-
-</project>
