+++ ci-droid-tasks-consumer-starter/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,1 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.societegenerale.cidroid.tasks.consumer.autoconfigure.CiDroidTasksConsumerAutoConfiguration
